requirements.txt
fastapi
uvicorn
google-genai
reportlab
python-multipart
PyPDF2